Add helper for adding LUKS keys, remove temp files
This is based in large part on suggestions and recommendations
from [Maryse47](https://github.com/Maryse47).

Instead of relying on temporary files, after this commit we rely
on piping around secrets almost exclusively. In places where it
is necessary to use a temporary file - for example, in order to
add a new keyfile using cryptsetup - we make use of a temporary
loopback file system, encrypted with a single-use key that is
discarded immediately after filesystem creation.

This closes #1.